Dedicated registerHotkeys file
